Setup Lerna + React Base Code
[Problem]
Need to add the React website code that will actually be served.
However, also needs to be able to support a monorepo structure, in order
to keep everything together.

[Solution]
Setup Lerna to better handle having all of the packages in one repo.
Then setup the existing CDK code and a new React website with it.

[Testing]
Confirmed both can be built and that their unit tests run fine.