Setup Simple S3 Bucket
[Problem]
Website needs to actually be deployed to an S3 bucket.

[Solution]
Updated the website stack to create an S3 bucket and deploy the React
website to it. Also updated the README with a boostrap command for the
CDK package.

[Testing]
Confirmed that the website is properly deployed and updated the unit
tests.