Setup API Gateway
[Problem]
Right now, the website doesn't actually use the Lambda to determine if a
login is valid.

[Solution]
Setup an API Gateway entry point for the website to reach out to.
Updaetd the website to actually use this endpoint and determine whether
the login was valid from it.

[Testing]
Manually confirmed everything works end-to-end.