Add Vaccination Support
[Problem]
Need to add support for users to set their vaccination status.
Originally planned to have an upload form, switching to self report, as
we should be able to trust everyone coming to our wedding (or they
wouldn't be invited :) )

[Solution]
Added support for a dropdown for vaccionation status. Will deprecate the
old vaccination form in the next commit.

[Testing]
Manually tested and confirmed working.